// backend/src/modules/foods/dto/market-price.dto.ts
export class MarketPriceDto {
    anio: string;
    mes: string;
    semana: string;
    fechaInicio: string;
    fechaTermino: string;
    region: string;
    sector: string;
    tipoPuntoMonitoreo: string;
    grupo: string;
    producto: string;
    unidad: string;
    precioMinimo: number;
    precioMaximo: number;
    precioPromedio: number;
}